requirements.txt
django
djangorestframework
djangorestframework-jsonapi
simplejson
